except stackoverflow, other datasets are ready!

+++ fedml_api/data_preprocessing/fed_cifar100/data_loader.py
@@ -28,9 +28,9 @@ def get_dataloader(dataset, data_dir, train_bs, test_bs, client_idx = None):
     train_h5 = h5py.File(train_file_path, 'r')
     test_h5 = h5py.File(test_file_path,'r')
     train_x, train_y, train_id = train_h5['image'], train_h5['label'], train_h5['id']
-    train_x = np.moveaxis(train_x[:], -1, 1)
+    train_x = np.moveaxis(train_x[:].astype(float), -1, 1)
     test_x, test_y, test_id = test_h5['image'], test_h5['label'], test_h5['id']
-    test_x = np.moveaxis(test_x[:], -1, 1)
+    test_x = np.moveaxis(test_x[:].astype(float), -1, 1)
     
     if client_idx is None:
         train_ds = data.TensorDataset(torch.tensor(train_x[:,:]), torch.tensor(train_y[:]))
